Refactor "cost" table into separate component
